+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/LandlordRegistrationJourney.kt
@@ -19,17 +19,17 @@ import uk.gov.communities.prsdb.webapp.forms.steps.Step
 import uk.gov.communities.prsdb.webapp.helpers.LandlordRegistrationJourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
 import uk.gov.communities.prsdb.webapp.models.emailModels.LandlordRegistrationConfirmationEmail
-import uk.gov.communities.prsdb.webapp.models.formModels.CheckAnswersFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.CountryOfResidenceFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.DateOfBirthFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.DeclarationFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.EmailFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.LookupAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.ManualAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.NameFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.NonEnglandOrWalesAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.PhoneNumberFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.SelectAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.CheckAnswersFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.CountryOfResidenceFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.DateOfBirthFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.DeclarationFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.EmailFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.LookupAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.ManualAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NameFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NonEnglandOrWalesAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.PhoneNumberFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.SelectAddressFormModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.CheckboxViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.RadiosButtonViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.SelectViewModel
