+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/PropertyRegistrationJourney.kt
@@ -21,20 +21,20 @@ import uk.gov.communities.prsdb.webapp.forms.steps.Step
 import uk.gov.communities.prsdb.webapp.helpers.PropertyRegistrationJourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
 import uk.gov.communities.prsdb.webapp.models.emailModels.PropertyRegistrationConfirmationEmail
-import uk.gov.communities.prsdb.webapp.models.formModels.DeclarationFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.HmoAdditionalLicenceFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.HmoMandatoryLicenceFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.LicensingTypeFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.LookupAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.ManualAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.NoInputFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.NumberOfHouseholdsFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.NumberOfPeopleFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.OccupancyFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.OwnershipTypeFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.PropertyTypeFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.SelectAddressFormModel
-import uk.gov.communities.prsdb.webapp.models.formModels.SelectiveLicenceFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.DeclarationFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.HmoAdditionalLicenceFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.HmoMandatoryLicenceFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.LicensingTypeFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.LookupAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.ManualAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NoInputFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NumberOfHouseholdsFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NumberOfPeopleFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.OccupancyFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.OwnershipTypeFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.PropertyTypeFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.SelectAddressFormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.SelectiveLicenceFormModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.CheckboxViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.RadiosButtonViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.RadiosDividerViewModel
