--- src/main/kotlin/uk/gov/communities/prsdb/webapp/models/formModels/CheckAnswersFormModel.kt
@@ -1,3 +1,0 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
-
-class CheckAnswersFormModel : FormModel
