+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/EmailFormModel.kt
@@ -1,25 +1,24 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
+import uk.gov.communities.prsdb.webapp.validation.EmailConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
-import uk.gov.communities.prsdb.webapp.validation.LengthConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.NotBlankConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class NonEnglandOrWalesAddressFormModel : FormModel {
+class EmailFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.nonEnglandOrWalesAddress.error.missing",
+                messageKey = "forms.email.error.missing",
                 validatorType = NotBlankConstraintValidator::class,
             ),
             ConstraintDescriptor(
-                messageKey = "forms.nonEnglandOrWalesAddress.error.tooLong",
-                validatorType = LengthConstraintValidator::class,
-                validatorArgs = arrayOf("0", "1000"),
+                messageKey = "forms.email.error.invalidFormat",
+                validatorType = EmailConstraintValidator::class,
             ),
         ],
     )
-    var nonEnglandOrWalesAddress: String = ""
+    var emailAddress: String? = null
 }
