+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/FormModel.kt
@@ -1,0 +1,3 @@
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
+
+interface FormModel
