+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/HmoAdditionalLicenceFormModel.kt
@@ -1,24 +1,25 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
+import uk.gov.communities.prsdb.webapp.validation.LengthConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.NotBlankConstraintValidator
-import uk.gov.communities.prsdb.webapp.validation.PositiveOrZeroIntegerValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class NumberOfPeopleFormModel : FormModel {
+class HmoAdditionalLicenceFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.numberOfPeople.input.error.missing",
+                messageKey = "forms.hmoAdditionalLicence.error.missing",
                 validatorType = NotBlankConstraintValidator::class,
             ),
             ConstraintDescriptor(
-                messageKey = "forms.numberOfPeople.input.error.invalidFormat",
-                validatorType = PositiveOrZeroIntegerValidator::class,
+                messageKey = "forms.licenceNumber.error.tooLong",
+                validatorType = LengthConstraintValidator::class,
+                validatorArgs = arrayOf("0", "255"),
             ),
         ],
     )
-    var numberOfPeople: String = ""
+    var licenceNumber: String? = null
 }
