+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/HmoMandatoryLicenceFormModel.kt
@@ -1,24 +1,25 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
+import uk.gov.communities.prsdb.webapp.validation.LengthConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.NotBlankConstraintValidator
-import uk.gov.communities.prsdb.webapp.validation.PhoneNumberConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class PhoneNumberFormModel : FormModel {
+class HmoMandatoryLicenceFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.phoneNumber.error.missing",
+                messageKey = "forms.hmoMandatoryLicence.error.missing",
                 validatorType = NotBlankConstraintValidator::class,
             ),
             ConstraintDescriptor(
-                messageKey = "forms.phoneNumber.error.invalidFormat",
-                validatorType = PhoneNumberConstraintValidator::class,
+                messageKey = "forms.licenceNumber.error.tooLong",
+                validatorType = LengthConstraintValidator::class,
+                validatorArgs = arrayOf("0", "255"),
             ),
         ],
     )
-    var phoneNumber: String? = null
+    var licenceNumber: String? = null
 }
