+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/LookupAddressFormModel.kt
@@ -1,25 +1,29 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
-import uk.gov.communities.prsdb.webapp.validation.LengthConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.NotBlankConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class HmoAdditionalLicenceFormModel : FormModel {
+class LookupAddressFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.hmoAdditionalLicence.error.missing",
+                messageKey = "forms.lookupAddress.postcode.error.missing",
                 validatorType = NotBlankConstraintValidator::class,
-            ),
-            ConstraintDescriptor(
-                messageKey = "forms.licenceNumber.error.tooLong",
-                validatorType = LengthConstraintValidator::class,
-                validatorArgs = arrayOf("0", "255"),
             ),
         ],
     )
-    var licenceNumber: String? = null
+    var postcode: String? = null
+
+    @ValidatedBy(
+        constraints = [
+            ConstraintDescriptor(
+                messageKey = "forms.lookupAddress.houseNameOrNumber.error.missing",
+                validatorType = NotBlankConstraintValidator::class,
+            ),
+        ],
+    )
+    var houseNameOrNumber: String? = null
 }
