+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/NameFormModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
