+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/NoInputFormModel.kt
@@ -1,0 +1,3 @@
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
+
+class NoInputFormModel : FormModel
