+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/OccupancyFormModel.kt
@@ -1,20 +1,19 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
-import uk.gov.communities.prsdb.webapp.constants.enums.LicensingType
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
 import uk.gov.communities.prsdb.webapp.validation.NotNullConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class LicensingTypeFormModel : FormModel {
+class OccupancyFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.licensingType.radios.error.missing",
+                messageKey = "forms.occupancy.radios.error.missing",
                 validatorType = NotNullConstraintValidator::class,
             ),
         ],
     )
-    var licensingType: LicensingType? = null
+    var occupied: Boolean? = null
 }
