+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/OwnershipTypeFormModel.kt
@@ -1,19 +1,20 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
+import uk.gov.communities.prsdb.webapp.constants.enums.OwnershipType
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
 import uk.gov.communities.prsdb.webapp.validation.NotNullConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class SelectAddressFormModel : FormModel {
+class OwnershipTypeFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.selectAddress.error.missing",
+                messageKey = "forms.ownershipType.radios.error.missing",
                 validatorType = NotNullConstraintValidator::class,
             ),
         ],
     )
-    var address: String? = null
+    var ownershipType: OwnershipType? = null
 }
