+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/PropertyTypeFormModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.constants.enums.PropertyType
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
@@ -8,7 +8,7 @@ import uk.gov.communities.prsdb.webapp.validation.NotNullConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class PropertyTypeFormModel : FormModel {
+class PropertyTypeFormModel : uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
