+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/requestModels/formModels/SelectAddressFormModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.formModels
+package uk.gov.communities.prsdb.webapp.models.requestModels.formModels
 
 import uk.gov.communities.prsdb.webapp.validation.ConstraintDescriptor
 import uk.gov.communities.prsdb.webapp.validation.IsValidPrioritised
@@ -6,14 +6,14 @@ import uk.gov.communities.prsdb.webapp.validation.NotNullConstraintValidator
 import uk.gov.communities.prsdb.webapp.validation.ValidatedBy
 
 @IsValidPrioritised
-class SelectLocalAuthorityFormModel : FormModel {
+class SelectAddressFormModel : FormModel {
     @ValidatedBy(
         constraints = [
             ConstraintDescriptor(
-                messageKey = "forms.selectLocalAuthority.error.missing",
+                messageKey = "forms.selectAddress.error.missing",
                 validatorType = NotNullConstraintValidator::class,
             ),
         ],
     )
-    var localAuthorityId: Int? = null
+    var address: String? = null
 }
