+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/forms/PageTests.kt
@@ -13,7 +13,7 @@ import org.springframework.validation.support.BindingAwareModelMap
 import uk.gov.communities.prsdb.webapp.exceptions.PrsdbWebException
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
 import uk.gov.communities.prsdb.webapp.forms.pages.Page
-import uk.gov.communities.prsdb.webapp.models.formModels.FormModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel
 import kotlin.test.Test
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
